improve API examples
