--- example/api/src/foobar.ts
@@ -1,9 +1,0 @@
-export async function oi(){
-  return "foobar";
-}
-
-export async function aye(){
-  return "foobar";
-}
-
-export * from "./woobar";
