--- example/api/src/woobar.ts
@@ -1,3 +1,0 @@
-export function bob(){
-  return "bob";
-}
