--- example/api/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "rootDir": "src/",
-    "outDir": "lib/",
-    "lib": ["es2018"],
-    "composite": true
-  }
-}
